change: modified boolean option parsing, don't parse the next value for boolean options.
Prevent the annoying prevous behavior when parsing a boolean option followed by variadic arguments parsed the first followed as value for option instead of true and parst the variadic arguments from the second element only. Like `--verbose first second` was parsed where the `verbose` value was `first` instead `true` and the followed array is only `[second]` instead `[first, second]`.

In order to achieve this the `shouldTakeNextAsValue` function extracted from `handleConcatenatedOpts` and use it in `handleOptWithoutValue` as well.

Resolves #187